+++ CHANGELOG.md
@@ -15,10 +15,11 @@ that can be found in the LICENSE file. -->
 - Fix filter option group. (#919)
 - Fix `originFileWithSubtype` and `fileWithSubtype` for livePhoto.
 - Fix: support only add permission for iOS/macOS. (#944)
+- Fix: modified the output path for iOS(add id in next path).
 
 ## 2.6.0
 
-### Features:
+### Features
 
 - Support `CustomFilter` for more filter options. (#901)
 - Add two new static methods for `PhotoManager`: